Rename the experiments list file and delete remaining redux refs
