+++ mlflow/server/js/src/experiment-tracking/reducers/Reducers.test.ts
@@ -36,7 +36,6 @@ import {
 import { mockExperiment, mockRunInfo } from '../utils/test-utils/ReduxStoreFixtures';
 import { RunTag, Param, ExperimentTag } from '../sdk/MlflowMessages';
 import {
-  SEARCH_EXPERIMENTS_API,
   GET_EXPERIMENT_API,
   GET_RUN_API,
   SEARCH_RUNS_API,
@@ -55,49 +54,6 @@ import { Stages, ModelVersionStatus } from '../../model-registry/constants';
 describe('test experimentsById', () => {
   test('should set up initial state correctly', () => {
     expect(experimentsById(undefined, {})).toEqual({});
-  });
-  test('searchExperiments correctly updates empty state', () => {
-    const experimentA = mockExperiment('experiment01', 'experimentA');
-    const experimentB = mockExperiment('experiment02', 'experimentB');
-    const state = undefined;
-    const action = {
-      type: fulfilled(SEARCH_EXPERIMENTS_API),
-      payload: {
-        experiments: [experimentA, experimentB],
-      },
-    };
-    const new_state = experimentsById(state, action);
-    expect(new_state).toEqual({
-      [experimentA.experimentId]: experimentA,
-      [experimentB.experimentId]: experimentB,
-    });
-  });
-  test('searchExperiments correctly updates state', () => {
-    const newA = mockExperiment('experiment01', 'experimentA');
-    const newB = mockExperiment('experiment02', 'experimentB');
-    const preserved = mockExperiment('experiment03', 'still exists');
-    const removed = mockExperiment('experiment04', 'removed');
-    const replacedOld = mockExperiment('experiment05', 'replacedOld');
-    const replacedNew = mockExperiment('experiment05', 'replacedNew');
-    const state = deepFreeze({
-      [preserved.experimentId]: preserved,
-      [removed.experimentId]: removed,
-      [replacedOld.experimentId]: replacedOld,
-    });
-    const action = {
-      type: fulfilled(SEARCH_EXPERIMENTS_API),
-      payload: {
-        experiments: [preserved, newA, newB, replacedNew],
-      },
-    };
-    const new_state = experimentsById(state, action);
-    expect(new_state).not.toEqual(state);
-    expect(new_state).toEqual({
-      [preserved.experimentId]: preserved,
-      [newA.experimentId]: newA,
-      [newB.experimentId]: newB,
-      [replacedNew.experimentId]: replacedNew,
-    });
   });
   test('getExperiment correctly updates empty state', () => {
     const experimentA = mockExperiment('experiment01', 'experimentA');
@@ -1385,27 +1341,6 @@ describe('test public accessors', () => {
       payload: payload,
     };
   }
-  test('Experiments', () => {
-    const A = {
-      experimentId: 'a',
-      name: 'A',
-      tags: [{ name: 'a', value: 'A' }, 'b'],
-    };
-    const B = {
-      experimentId: 'b',
-      name: 'B',
-    };
-    const action = new_action({
-      type: fulfilled(SEARCH_EXPERIMENTS_API),
-      payload: { experiments: [A, B] },
-    });
-    const state = rootReducer(undefined, action);
-    expect(state.entities.experimentTagsByExperimentId).toEqual({});
-    expect(getExperiments(state)).toEqual([A, B]);
-    expect(getExperiment(A.experimentId, state)).toEqual(A);
-    expect(getExperimentTags(B.experimentId, state)).toEqual({});
-    expect(getExperimentTags(A.experimentId, state)).toEqual({});
-  });
   test('tags, params and runinfo', () => {
     const key1 = 'key1';
     const key2 = 'key2';
