+++ mlflow/server/js/src/experiment-tracking/route-defs.ts
@@ -28,7 +28,7 @@ export const getRouteDefs = () => [
   {
     path: RoutePaths.experimentPageTabbed,
     element: createLazyRouteElement(() => {
-      return import(/* webpackChunkName: "experimentPage" */ './components/ExperimentPage');
+      return import(/* webpackChunkName: "experimentPage" */ './components/ExperimentListPage');
     }),
     pageId: 'mlflow.experiment.details.tab',
   },
@@ -45,21 +45,21 @@ export const getRouteDefs = () => [
   {
     path: RoutePaths.experimentPage,
     element: createLazyRouteElement(
-      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentPage'),
+      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentListPage'),
     ),
     pageId: 'mlflow.experiment.details',
   },
   {
     path: RoutePaths.experimentPageSearch,
     element: createLazyRouteElement(
-      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentPage'),
+      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentListPage'),
     ),
     pageId: 'mlflow.experiment.details.search',
   },
   {
     path: RoutePaths.compareExperimentsSearch,
     element: createLazyRouteElement(
-      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentPage'),
+      () => import(/* webpackChunkName: "experimentPage" */ './components/ExperimentListPage'),
     ),
     pageId: 'mlflow.experiment.compare',
   },
